Extract imperatives, acceptability, and manner adjustment
